requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
celery
redis
face_recognition
scikit-learn
python-dotenv
gunicorn
pymysql
PyJWT